backend/requirements.txt
Django==5.2.6
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
Pillow==10.0.1
psycopg2-binary==2.9.7

# Para IA y Machine Learning
azure-cognitiveservices-vision-face==0.6.0
azure-cognitiveservices-vision-computervision==0.9.0
opencv-python==4.8.1.78
numpy==1.24.3
scikit-learn==1.3.0
tensorflow==2.13.0
torch==2.0.1
torchvision==0.15.2

# Para procesamiento de imágenes y OCR
pytesseract==0.3.10
pdf2image==1.16.3

# Para notificaciones push
pyfcm==1.5.4

# Para manejo de archivos y medios
python-magic==0.4.27

# Para validaciones y utilidades
python-decouple==3.8
celery==5.3.2
redis==5.0.1

# Para desarrollo y testing
pytest==7.4.2
pytest-django==4.5.2
factory-boy==3.3.0
coverage==7.3.2

# Para documentación de API
drf-spectacular==0.26.5

# Para autenticación JWT
djangorestframework-simplejwt==5.3.0

# Para manejo de fechas
python-dateutil==2.8.2

# Para criptografía
cryptography==41.0.7

# Para archivos Excel/CSV
openpyxl==3.1.2
pandas==2.1.1

# Para envío de emails
django-ses==3.5.0

# Para almacenamiento en la nube (AWS S3)
boto3==1.28.57
django-storages==1.14

# Para logging y monitoreo
sentry-sdk==1.32.0

# Para cache
django-redis==5.4.0